==> R/wrangling/6_merge_habitat_spatial_covariates_broad.R <==
###
# Project: Parks - Abrolhos
# Data:    BOSS fish, habitat
# Task:    Modelling fish abundance w/ FSSGAM
# author:  Claude & Kingsley
# date:    Nov-Dec 2021
##

# Merge habitat and spatial covariates
rm(list=ls())

library(reshape2)
library(raster)
library(sp)
library(ggplot2)
library(dplyr)

#read in habitat
habitat <- read.csv("data/tidy/montebello.synthesis.complete.habitat.csv")%>%
  dplyr::mutate(mesophotic.reef = biota.crinoids+biota.invertebrate.complex+biota.octocoral.black+
           biota.sponges+biota.hydroids)%>%
  dplyr::mutate(photic.reef = biota.stony.corals + biota.macroalgae)%>%
  glimpse()

# get spatial covariates (made in 'R/wrangling/A1_spatial_layers.R')
tifs  <- list.files("output/spatial_covariates/", "*.tif", full.names = TRUE)
preds <- stack(tifs)
plot(preds)
## extract bathy derivatives for modelling
# spatial setup
wgscrs <- CRS("+proj=longlat +datum=WGS84 +south")
# utmcrs <- CRS("+proj=utm +zone=50 +south +datum=WGS84 +units=m +no_defs")

# align crs and check samples over bathy and extract terrain data
allhab_sp <- SpatialPointsDataFrame(coords = habitat[22:21], data = habitat, 
                                    proj4string = wgscrs)
# allhab_t  <- spTransform(allhab_sp, CRS = utmcrs)
plot(preds[[1]])
plot(allhab_sp, add=T)
habt_df   <- as.data.frame(allhab_sp, xy = T)
habi_df   <- cbind(habt_df, raster::extract(preds, allhab_sp))

#tidy names
habi_df <- habi_df %>%
  dplyr::rename("depth_ga" = layer_ga_Z, "detrended" = layer_ga_detrended, 
                "roughness" = layer_ga_roughness, "slope" = layer_ga_slope,
                "tpi" = layer_ga_tpi)%>%
  dplyr::mutate(depth_ga = abs(depth_ga))%>%
  glimpse()

names(habi_df)

#save out the habitat with spatial covariates
saveRDS(habi_df, "data/tidy/broad_merged_habitat.rds")
